refactor: move localdb pkg to dashboard
- its only used by the dashboard
